Combined LittleFS.h and LittleFS_NAND.h
