+++ src/LittleFS.h
@@ -25,6 +25,7 @@
 #include <FS.h>
 #include <SPI.h>
 #include "littlefs/lfs.h"
+//#include <algorithm>
 
 class LittleFSFile : public File
 {
@@ -519,7 +520,127 @@ public:
 };
 #endif
 
+class LittleFS_SPINAND : public LittleFS
+{
+public:
+	LittleFS_SPINAND() {
+		port = nullptr;
+	}
+	bool begin(uint8_t cspin, SPIClass &spiport=SPI);
+	uint8_t readECC(uint32_t address, uint8_t *data, int length);
+	void readBBLUT(uint16_t *LBA, uint16_t *PBA, uint8_t *linkStatus);
+	bool lowLevelFormat(char progressChar, Print* pr=&Serial);
+	uint8_t addBBLUT(uint32_t block_address);  //temporary for testing
+  
+private:
+	int read(lfs_block_t block, lfs_off_t offset, void *buf, lfs_size_t size);
+	int prog(lfs_block_t block, lfs_off_t offset, const void *buf, lfs_size_t size);
+	int erase(lfs_block_t block);
+	int wait(uint32_t microseconds);
+	static int static_read(const struct lfs_config *c, lfs_block_t block,
+	  lfs_off_t offset, void *buffer, lfs_size_t size) {
+		//Serial.printf("  flash rd: block=%d, offset=%d, size=%d\n", block, offset, size);
+		return ((LittleFS_SPINAND *)(c->context))->read(block, offset, buffer, size);
+	}
+	static int static_prog(const struct lfs_config *c, lfs_block_t block,
+	  lfs_off_t offset, const void *buffer, lfs_size_t size) {
+		//Serial.printf("  flash wr: block=%d, offset=%d, size=%d\n", block, offset, size);
+		return ((LittleFS_SPINAND *)(c->context))->prog(block, offset, buffer, size);
+	}
+	static int static_erase(const struct lfs_config *c, lfs_block_t block) {
+		//Serial.printf("  flash er: block=%d\n", block);
+		return ((LittleFS_SPINAND *)(c->context))->erase(block);
+	}
+	static int static_sync(const struct lfs_config *c) {
+		return 0;
+	}
+  bool isReady();
+  bool writeEnable();
+  void eraseSector(uint32_t address);
+  void writeStatusRegister(uint8_t reg, uint8_t data);
+  uint8_t readStatusRegister(uint16_t reg, bool dump);
+  void loadPage(uint32_t address);
 
+  void deviceReset();
+  
+	SPIClass *port;
+	uint8_t pin;
+	uint8_t addrbits;
+	uint32_t progtime;
+	uint32_t erasetime;
+	uint32_t chipsize;
+	uint32_t blocksize;
+	
+private:
+  uint8_t die = 0;      //die = 0: use first 1GB die PA[16], die = 1: use second 1GB die PA[16].
+  uint8_t dies = 0;		//used for W25M02
+  uint32_t capacityID ;   // capacity
+  uint32_t deviceID;
+
+  uint16_t eccSize = 64;
+  uint16_t PAGE_ECCSIZE = 2112;
+
+};
+
+
+#if defined(__IMXRT1062__)
+class LittleFS_QPINAND : public LittleFS
+{
+public:
+	LittleFS_QPINAND() { }
+	bool begin();
+    bool deviceErase();
+	uint8_t readECC(uint32_t targetPage, uint8_t *buf, int size);
+	void readBBLUT(uint16_t *LBA, uint16_t *PBA, uint8_t *linkStatus);
+	bool lowLevelFormat(char progressChar);
+	uint8_t addBBLUT(uint32_t block_address);  //temporary for testing
+	
+private:
+	int read(lfs_block_t block, lfs_off_t offset, void *buf, lfs_size_t size);
+	int prog(lfs_block_t block, lfs_off_t offset, const void *buf, lfs_size_t size);
+	int erase(lfs_block_t block);
+	int wait(uint32_t microseconds);
+	static int static_read(const struct lfs_config *c, lfs_block_t block,
+	  lfs_off_t offset, void *buffer, lfs_size_t size) {
+		//Serial.printf(".....  flash rd: block=%d, offset=%d, size=%d\n", block, offset, size);
+		return ((LittleFS_QPINAND *)(c->context))->read(block, offset, buffer, size);
+	}
+	static int static_prog(const struct lfs_config *c, lfs_block_t block,
+	  lfs_off_t offset, const void *buffer, lfs_size_t size) {
+		//Serial.printf(".....  flash wr: block=%d, offset=%d, size=%d\n", block, offset, size);
+		return ((LittleFS_QPINAND *)(c->context))->prog(block, offset, buffer, size);
+	}
+	static int static_erase(const struct lfs_config *c, lfs_block_t block) {
+		//Serial.printf(".....  flash er: block=%d\n", block);
+		return ((LittleFS_QPINAND *)(c->context))->erase(block);
+	}
+	static int static_sync(const struct lfs_config *c) {
+		return 0;
+	}
+  bool isReady();
+  bool writeEnable();
+  void deviceReset();
+  void eraseSector(uint32_t address);
+  void writeStatusRegister(uint8_t reg, uint8_t data);
+  uint8_t readStatusRegister(uint16_t reg, bool dump);
+  
+	uint8_t addrbits;
+	uint32_t progtime;
+	uint32_t erasetime;
+	uint32_t chipsize;
+	uint32_t blocksize;
+	
+private:
+  uint8_t die = 0;      //die = 0: use first 1GB die, die = 1: use second 1GB die.
+  uint8_t dies;
+  uint32_t capacityID ;   // capacity
+  uint32_t deviceID;
+
+  uint16_t eccSize = 64;
+  uint16_t PAGE_ECCSIZE = 2112;
+
+};
+#endif
